+++ tests/testthat/test-template-content.R
@@ -16,8 +16,8 @@ test_that("og tags are populated on index.html", {
 })
 
 test_that("og tags are populated on reference pages", {
-  pork_html <- read_lines(path(pkg$dst_path, "reference", "pulledpork.html"))
-  desc <- '<meta property="og:description" content="Pulled pork is delicious" />'
+  pork_html <- read_lines(path(pkg$dst_path, "reference", "f.html"))
+  desc <- '<meta property="og:description" content="Title" />'
   expect_true(desc %in% pork_html)
   img <- '<meta property="og:image" content="http://example.com/pkg/logo.png" />'
   expect_true(img %in% pork_html)